Refactor OrganizationComponent series
